backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
chromadb
torch
torchvision
transformers
pillow
sentence-transformers
pydantic
